+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 1.1.0 (2018-05-29)
 * Add vcs modules support in format: "vcs+protocol://repo_url/@vcsVersion#egg=pkg-pkgVersion" (#2)
 * Support module features syntax: module\[feature]:version (#3)
 * PipModule api changes: